Refactor test file
Moved out base classes and left only test cases
